backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
python-dotenv==1.2.2
supabase==2.28.3
langgraph==1.1.6
